Identify certificates by issuer and subject
When a file contains multiple certificates, do not use the first line
of the certificate in the file to identify it since a preceding
certificate may be longer or shorter when renewed.

Instead, use the certificate issuer and subject to identify the
certificate.

While here, add the first line as an extra attribute of the generated
event.